// frontend/src/app/features/root-finder/models/parameters.ts
export type IntervalParameters = {
  lowerBound: string;
  upperBound: string;
};

export type OneGuessParameters = {
  guess: string;
};

export type OneGuessWithMultiplicityParameters = {
  guess: string;
  multiplicity: number;
};

export type TwoGuessesParameters = {
  firstGuess: string;
  secondGuess: string;
};